refactor: extract portal_theme domain into ferriskey-portal-theme
