+++ libs/ferriskey-portal-theme/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "ferriskey-portal-theme"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-domain = { path = "../ferriskey-domain" }
+chrono = { version = "0.4.41", features = ["serde"] }
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.141"
+utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
+uuid = { version = "1.16.0", features = ["serde", "v4", "v7"] }
+mockall = { version = "0.14.0", optional = true }
+
+[dev-dependencies]
+mockall = "0.14.0"
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros"] }
+ferriskey-domain = { path = "../ferriskey-domain", features = ["mock"] }
+
+[features]
+mock = ["mockall"]
