+++ libs/ferriskey-portal-theme/src/policies.rs
@@ -1,15 +1,12 @@
-use crate::domain::{
-    authentication::value_objects::Identity,
-    client::ports::ClientRepository,
-    common::{
-        entities::app_errors::CoreError,
-        policies::{FerriskeyPolicy, Policy},
-    },
-    portal_theme::ports::PortalThemePolicy,
-    realm::entities::Realm,
-    role::entities::permission::Permissions,
-    user::ports::{UserRepository, UserRoleRepository},
-};
+use ferriskey_domain::auth::Identity;
+use ferriskey_domain::client::ports::ClientRepository;
+use ferriskey_domain::common::app_errors::CoreError;
+use ferriskey_domain::common::policies::{FerriskeyPolicy, Policy};
+use ferriskey_domain::realm::Realm;
+use ferriskey_domain::role::permission::Permissions;
+use ferriskey_domain::user::ports::{UserRepository, UserRoleRepository};
+
+use crate::ports::PortalThemePolicy;
 
 impl<U, C, UR> PortalThemePolicy for FerriskeyPolicy<U, C, UR>
 where
